# feat: support message transformer in the context of "adopt" (fixes #60)