--- a/src/Roarr.ts
+++ b/src/Roarr.ts
@@ -69,7 +69,7 @@ export type {
   Message,
   MessageEventHandler,
   RoarrGlobalState,
-  TranslateMessageFunction,
+  TransformMessageFunction,
 } from './types';
 
 export {