--- a/src/browser.ts
+++ b/src/browser.ts
@@ -35,7 +35,7 @@ export type {
   Message,
   MessageEventHandler,
   RoarrGlobalState,
-  TranslateMessageFunction,
+  TransformMessageFunction,
 } from './types';
 
 export {