--- a/src/utilities/index.ts
+++ b/src/utilities/index.ts
@@ -1,0 +1,3 @@
+export {
+  hasOwnProperty,
+} from './hasOwnProperty';