--- a/test/roarr/roarr.ts
+++ b/test/roarr/roarr.ts
@@ -70,6 +70,17 @@ test('logs an empty message when first parameter is an object and the second par
   ]);
 });
 
+test('throws in case of invalid invocation', (t) => {
+  const log = createLoggerWithHistory();
+
+  t.throws(() => {
+    // @ts-expect-error Invalid invocation
+    log({}, {});
+  }, {
+    message: 'Message must be a string. Received object.',
+  });
+});
+
 test('formats message using sprintf', (t) => {
   const log = createLoggerWithHistory();
 
@@ -370,7 +381,7 @@ test('throws an error if child does not return an object', (t) => {
         return '';
       })('foo');
   }, {
-    message: 'Child middleware function must return a message object.',
+    message: 'Message transform function must return a message object.',
   });
 });