Before:

```
simple message x 1,456,083 ops/sec ±0.99% (95 runs sampled)
message with printf x 1,092,793 ops/sec ±0.95% (96 runs sampled)
message with context x 1,307,103 ops/sec ±0.98% (95 runs sampled)
message with large context x 1,399,549 ops/sec ±1.72% (75 runs sampled)
message with large context x 1,387,441 ops/sec ±1.92% (75 runs sampled)
```

After:

```
simple message x 1,309,177 ops/sec ±0.67% (97 runs sampled)
message with printf x 969,862 ops/sec ±0.94% (97 runs sampled)
message with context x 1,112,226 ops/sec ±1.04% (98 runs sampled)
message with large context x 1,231,458 ops/sec ±1.75% (76 runs sampled)
message with large context x 1,245,710 ops/sec ±1.59% (76 runs sampled)
```

Insignificant performance degradation.