@gajus 2022-03-18
:tada: This PR is included in version 7.10.0 :tada:

The release is available on:
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/roarr/v/7.10.0)
- [GitHub release](https://github.com/gajus/roarr/releases/tag/v7.10.0)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: